wip(basic): #9 refactor to allow alternative configurations
